Fix summarize history recipe by using the workspace root as the git root.
