+++ client/cody-shared/src/chat/recipes/git-log.ts
@@ -1,5 +1,4 @@
 import { spawnSync } from 'child_process'
-import * as path from 'path'
 
 import { CodebaseContext } from '../../codebase-context'
 import { Editor } from '../../editor'
@@ -22,11 +21,10 @@ export class GitHistory implements Recipe {
         _intentDetector: IntentDetector,
         _codebaseContext: CodebaseContext
     ): Promise<Interaction | null> {
-        const activeEditor = editor.getActiveTextEditor()
-        if (!activeEditor) {
+        const dirPath = editor.getWorkspaceRootPath()
+        if (!dirPath) {
             return null
         }
-        const dirPath = path.dirname(activeEditor.filePath)
 
         const logFormat = '--pretty="Commit author: %an%nCommit message: %s%nChange description:%b%n"'
         const items = [
@@ -60,15 +58,17 @@ export class GitHistory implements Recipe {
         const gitLogOutput = gitLogCommand.stdout.toString().trim()
 
         if (!gitLogOutput) {
-            // TODO: Show the warning within the Chat UI.
-            // editor.showWarningMessage('No git history found for the selected option.')
+            console.error(
+                'No git history found for the selected option.',
+                gitLogCommand.stderr.toString(),
+                gitLogCommand.error?.message
+            )
             return null
         }
 
         const truncatedGitLogOutput = truncateText(gitLogOutput, MAX_RECIPE_INPUT_TOKENS)
         if (truncatedGitLogOutput.length < gitLogOutput.length) {
-            // TODO: Show the warning within the Chat UI.
-            // editor.showWarningMessage('Truncated extra long git log output, so summary may be incomplete.')
+            console.warn('Truncated extra long git log output, so summary may be incomplete.')
         }
 
         const timestamp = getShortTimestamp()
