+++ client/cody/src/editor/vscode-editor.ts
@@ -10,6 +10,21 @@ import {
 const SURROUNDING_LINES = 50
 
 export class VSCodeEditor implements Editor {
+    public getWorkspaceRootPath(): string | null {
+        const uri = vscode.window.activeTextEditor?.document.uri
+        if (uri) {
+            const wsFolder = vscode.workspace.getWorkspaceFolder(uri)
+            if (wsFolder) {
+                return wsFolder.uri.fsPath
+            }
+        }
+
+        if (vscode.workspace.workspaceFolders && vscode.workspace.workspaceFolders.length >= 1) {
+            return vscode.workspace.workspaceFolders[0].uri.fsPath
+        }
+        return null
+    }
+
     public getActiveTextEditor(): ActiveTextEditor | null {
         const activeEditor = vscode.window.activeTextEditor
         if (!activeEditor || activeEditor.document.uri.scheme !== 'file') {
